opt changes added for login and sign up

+++ src/app/otp/otp.page.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { ActivatedRoute } from '@angular/router';
+import { getAuth } from 'firebase/auth';
+import firebase from '@firebase/app-compat';
+import { ToastController } from '@ionic/angular';
 @Component({
   selector: 'app-otp',
   templateUrl: './otp.page.html',
@@ -11,8 +14,15 @@ export class OtpPage implements OnInit {
   otp: string;
   username: string = '';
   phoneNumber: string = '';
+  reCaptchaVerifier;
+  auth = getAuth();
+  otpConfirmation: firebase.default.auth.ConfirmationResult;
 
-  constructor(private router: Router, private route: ActivatedRoute) {
+  constructor(
+    private router: Router,
+    private route: ActivatedRoute,
+    public toastController: ToastController
+  ) {
     this.route.queryParams.subscribe((params) => {
       this.from = params.from;
       this.username = params.username;
@@ -22,29 +32,52 @@ export class OtpPage implements OnInit {
 
   ngOnInit() {}
 
-  createPassword() {
-    // this.otpConfirmation.confirm(this.otp).then((result) => {
-    //   // User signed in successfully.
-    //   console.log(result);
-    //   const user = result.user;
-    //   // ...
-    // }).catch((error) => {
-    //   // User couldn't sign in (bad verification code?)
-    //   // ...
-    // });
-    if (this.from == 'create') {
-      this.router.navigate(['/', 'app', 'create-password'], {
-        queryParams: { username: this.username, phoneNumber: this.phoneNumber },
+  ionViewDidEnter() {
+    this.reCaptchaVerifier = new firebase.auth.RecaptchaVerifier(
+      'recaptcha-container'
+    );
+    firebase
+      .auth()
+      .signInWithPhoneNumber('+91' + this.phoneNumber, this.reCaptchaVerifier)
+      .then((data) => {
+        this.otpConfirmation = data;
+      })
+      .catch((err) => {
+        console.log(err);
       });
-    } else if (this.from == 'login') {
-      localStorage.setItem("isLinkInBioLoggedIn","yes")
-      this.router.navigate(['/', 'vivek'], {
-        queryParams: { username: this.username },
+  }
+
+  async createPassword() {
+    try {
+      this.otpConfirmation
+        .confirm(this.otp)
+        .then((result) => {
+          console.log(result);
+          if (this.from == 'create') {
+            this.router.navigate(['/', 'app', 'create-password'], {
+              queryParams: {
+                username: this.username,
+                phoneNumber: this.phoneNumber,
+              },
+            });
+          } else if (this.from == 'login') {
+            localStorage.setItem('isLinkInBioLoggedIn', 'yes');
+            this.router.navigate(['/', this.username], {
+              queryParams: { username: this.username },
+            });
+          }
+        })
+        .catch((error) => {
+          console.log(error);
+        });
+    } catch (error) {
+      const toast = await this.toastController.create({
+        message: 'OTP error ',
+        duration: 2000,
       });
+      toast.present();
     }
   }
 
-  ionViewDidLeave(){
-    
-  }
+  ionViewDidLeave() {}
 }
